first working code

+++ src/main/kotlin/eu/wojciechzurek/mattermostnbp/api/Rate.kt
@@ -1,8 +1,10 @@
 package eu.wojciechzurek.mattermostnbp.api
 
+import java.math.BigDecimal
+
 data class Rate(
         val currency: String,
-        val mid: Float? = null,
-        val ask: Float? = null,
-        val bid: Float? = null
+        val mid: BigDecimal? = null,
+        val ask: BigDecimal? = null,
+        val bid: BigDecimal? = null
 )
