+++ src/main/kotlin/eu/wojciechzurek/mattermostnbp/api/TableAResponse.kt
@@ -1,6 +1,7 @@
 package eu.wojciechzurek.mattermostnbp.api
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
+import java.math.BigDecimal
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class TableAResponse(
@@ -11,5 +12,5 @@ data class TableAResponse(
 data class TableA(
         val code: String,
         val currency: String,
-        val mid: Float
+        val mid: BigDecimal
 )
